Load Delayed::JobWrapper before ActiveJob is loaded
